backend/requirements.txt
torch
fastapi
uvicorn[standard]
pydantic
numpy
google-generativeai